requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
blinker==1.8.2
blis==0.7.11
boto3==1.34.121
botocore==1.34.121
cachetools==4.2.4
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.6
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==5.9
gast==0.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==2.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.13.1
Keras-Preprocessing==1.1.2
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langcodes==3.4.0
langsmith==0.1.75
language-data==1.2.0
libclang==18.1.1
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.28.0
opt-einsum==3.3.0
orjson==3.10.3
packaging==23.2
pandas==2.0.3
pdfreader==0.1.15
pillow==10.3.0
pkgutil-resolve-name==1.3.10
plac==1.1.3
preshed==3.0.9
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycryptodome==3.20.0
pydantic==2.7.3
pydantic-core==2.18.4
pydeck==0.9.1
pygments==2.18.0
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytokenizations==0.8.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
SQLAlchemy==2.0.30
srsly==2.4.8
streamlit==1.34.0
streamlit-modal==0.1.2
sympy==1.12.1
tenacity==8.3.0
tensorboard==2.2.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.2.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thinc==8.2.4
threadpoolctl==3.5.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchcontrib==0.0.2
tornado==6.4.1
tqdm==4.66.4
transformers==4.36.2
typer==0.12.3
typing-extensions==4.12.1
typing-inspect==0.9.0
tzdata==2024.1
urllib3==1.26.18
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.4.1
werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.19.2
